qweb/djangobd/requirements.txt
amazon-braket-default-simulator==1.6.2
amazon-braket-ocean-plugin==1.0.10
amazon-braket-schemas==1.10.0
amazon-braket-sdk==1.25.2
antlr4-python3-runtime==4.11.1
attrs==23.2.0
backoff==2.2.1
boltons==23.1.1
boto3==1.24.20
botocore==1.27.20
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.12
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
connexion==2.13.1
cryptography==42.0.1
dill==0.3.7
dimod==0.12.14
diskcache==5.6.3
docker==7.0.0
docplex==2.25.236
dwave-cloud-client==0.10.6
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SSLify==0.1.5
homebase==1.0.1
idna==3.6
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonref==1.1.0
jsonschema==4.6.0
MarkupSafe==2.1.4
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==2.8.4
ninja==1.11.1
numpy==1.26.3
openapi==1.1.0
openqasm3==0.4.0
opt-einsum==3.3.0
packaging==23.2
pbr==6.0.0
plucky==0.4.3
ply==3.11
psutil==5.9.8
pycparser==2.21
pydantic==1.10.14
pyrsistent==0.20.0
PySocks==1.7.1
pyspnego==0.10.2
python-constraint==1.4.0
python-dateutil==2.8.2
PyYAML==6.0.1
qiskit==1.0.1
qiskit-aer==0.13.3
qiskit-ibm-provider==0.10.0
qiskit-optimization==0.6.1
qiskit-qasm3-import==0.4.1
requests==2.28.0
requests-ntlm==1.2.0
retworkx==0.14.0
rustworkx==0.14.0
s3transfer==0.6.2
scipy==1.12.0
six==1.16.0
sspilib==0.1.0
stevedore==5.1.0
swagger-ui-bundle==0.0.9
symengine==0.11.0
sympy==1.12
tweedledum==1.1.1
typing_extensions==4.9.0
urllib3==1.26.9
websocket-client==1.7.0
websockets==12.0
Werkzeug==2.3.4